services/sql-bot/requirements.txt
fastapi
uvicorn
httpx
pydantic
# chromadb used to assist with schema search but replaced by RiskSQLSchema
langgraph
google-genai
langchain
langsmith
psycopg2-binary